adapt eth logic to Fusaka EIP-7594 version & fix synchronizer crafted_items_root initialization & get crafted_item mt-proof (#11)
* adapt eth logic to Fusaka EIP-7594 version ( https://blog.ethereum.org/2025/10/15/fusaka-blob-update )

At the current date (2025-10-23):
- Before this commit, the code worked fine with infura.io nodes but it failed with publicnode.com nodes.
- With this commit, the code works fine whith publicnode.com nodes but it fails with infura.io nodes.

Also adds the full-flow.sh script to update the config file to start
from the latest blob slot, and to run the tmux pannels with the
synchronizer and app.

* update synchronizer to initialize the crafted_items_root to avoid error when processing first blob

initialize the `created_items_root` with 0x00... root, so that when new
items are crafted from scratch, their `payload.created_items_root`
(which is 0x00... since it a from-scratch item) is accepted as a "valid"
one, since it ears at the `created_items_root`.

* add app request to synchronizer asking for the merkleproof of the item in the created_items